+++ CHANGELOG.md
@@ -44,7 +44,8 @@ New features:
 + As a *Servlet* in the HRSPortlets web application, adds `/go` redirector that takes a `urlKey`
   request parameter. When this maps to a URL known to the HrsUrlsDao, redirects to that URL. When
   this does not map to a URL known to that DAO, responds 404 not found. ([#126][])
-
++ Adds a JSON resource URL for asking whether the user has a specific HRS Portlets role. Intended
+  for use in `switch` widget type to switch widget behavior on whether user has role. ([ #127][])
 
 ### 3.1.0: targeted notifications and notices for PHIT
 
@@ -398,6 +399,7 @@ This and many more earlier releases exist as [releases in the GitHub repo][].
 [#123]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/123
 [#125]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/125
 [#126]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/126
+[#127]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/127
 [#128]: https://github.com/UW-Madison-DoIT/hrs-portlets/pull/128
 
 [HRSPLT-346]: https://jira.doit.wisc.edu/jira/browse/HRSPLT-346